refactor(characters): extract Characters.Generation
Fourth slice of #134. random_decisions, root_concept_ids, and their
recursion helpers form the character-building decision generation
concern. The issue's proposed module list predates this being a
distinct seam; it fits none of Store/Effects/Leveling/Progression/
Preparation, so it gets its own module rather than being shoehorned in.
